backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.0
drf-yasg==1.21.6
PyJWT==2.8.0
python-decouple==3.8
Pillow==12.2.0
